requirements.txt
requests>=2.25.1
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
nltk>=3.6.0
sentence-transformers>=2.0.0
torch>=1.9.0
transformers>=4.0.0
pyyaml>=5.4.0
tqdm>=4.60.0
jupyter>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0